notification_service/pyproject.toml
[project]
name = "notification_service"
version = "0.1.0"
description = "..."
authors = [{name = "Jahongir", email = "jahongir.hodjaniyazov@gmail.com"}]
requires-python = ">=3.13"
dependencies = [
    "aiokafka ==0.12.0",
    "fastapi ==0.116.1",
    "pydantic ==2.11.7",
    "pydantic-settings ==2.10.1",
    "sqlalchemy ==2.0.41",   
    "uvicorn ==0.35.0", 
    "asyncpg ==0.30.0",
    "python-jose ==3.5.0",
    "redis[hiredis] ==6.4.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build_backend = "setuptools.build_meta"